move helper scripts to their own folder
