--- Editor/JsonHelper.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 1f26b192186a48f5bd2fb938b1551d45
-timeCreated: 1563357184
