+++ Editor/ProblemDescriptorHelper.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.IO;
+using Packages.ProjectAuditor.Editor.Utils;
 
 namespace Unity.ProjectAuditor.Editor
 {
